appli/requirements.txt
Flask
pymongo
bcrypt
plotly
requests
kafka-python
geocoder==1.38.1
